more intuitive answers

+++ libc.h
@@ -1,8 +1,7 @@
 #include <errno.h>
-#include <stddef.h>
 
 #undef errno
 extern int errno;
 
 _Noreturn void _exit(int);
-int write(int, const void *, size_t);
+int write(int, const void *, unsigned);
